requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
Django==4.2.11
django-jazzmin==3.0.0
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
graphene==3.3
graphene-django==3.2.1
graphql-core==3.2.3
graphql-relay==3.2.0
inflection==0.5.1
packaging==24.0
pillow==10.3.0
promise==2.3
psycopg2-binary==2.9.9
PyJWT==1.7.1
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
typing_extensions==4.11.0
uritemplate==4.1.1
